Make require_durable_file_mounts server-authoritative; move tests to module home
